feat: Improve image resizing tool with hover effects and style adjustments
- Added opacity transition to ImageResizeTool for smoother visibility changes.
- Updated resize handle dimensions and colors for better visibility and user experience.
- Integrated hover state management in ImageResizeTool to enhance interaction feedback.
- Refined QuillNextImage to conditionally display the ImageResizeTool based on loading state, improving performance.
