+++ packages/website/src/components/AnimatedBackground.css
@@ -1,0 +1,13 @@
+.animated-background {
+  opacity: 0;
+  animation: fadeIn 0.8s ease-in forwards;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+  }
+  to {
+    opacity: 1;
+  }
+}
